refactor: move welcome-message.ts into lib/content directory
