+++ lib/content/__mocks__/welcome-message.ts
