+++ lib/content/welcome-message.ts
